Add unit tests for empty values in CSV
There is no difference between the following lines in CSV:
1) f1,"",f3
2) f1,,f3

Univocity parser returns {"f1", null, "f3"} for both rows. There is no way to tell between a missing value and an explicit empty string "".

We covered this situation with a unit test and also added an explicit code that always returns null instead of "" in case we change the CSV parser.