Cache certain field values to optimize memory usage
There are many fields that repeat for multiple rows of a single table, e.g., stop_headsign for stop_times.txt or shape_id for shapes.txt. Caching of those values dramatically saves memory for large feeds, e.g., a feed with 25 M stop times requires 8 GiB without caching and 3 GiB with caching.

All tables are read in parallel, that's why we create a separate set of caches for each table. It is tempting to use the same cache for, e.g., trip_id in both trips.txt and stop_times.txt but that would require synchronization that slows down the reading.

The following fields are cached:
* annotated with @CachedField;
* types ID, date, time, color.

Primary keys are not cached since they are unique for each row.