Use java.nio.file.Path and support loading of ZIP files to memory
java.nio.file.Path is a more flexible way to retrieve file and
directories.

ZIP file can be loaded to memory instead of being saved to a temporary
directory. This is useful when the ZIP is placed on a non-local file
system or has to be downloaded from the given URL.