fix: Do not crash on 0 entities in feed_info.txt
The previously used method isEmptyFile() tells if the whole text file is
empty. That crashed the Validator for a non-empty feed_info.txt file
that contains only the headers.

This commit adds documentation to isEmptyFile() and other methods in
GtfsTableContainer.

The unit test was changed to use the real TableContainer classes with
their real behaviour.