feat: Strip leading and trailing whitespaces and emit a warning
GTFS reference says:
Remove any extra spaces between fields or field names. Many parsers
consider the spaces to be part of the value, which may cause errors.

GTFS Validator strips whitespaces from protected values. We do not
see any use case when such a whitespace may be needed. On the other
hand, some real-world feeds use trailing whitespaces for some values
and omit them for the others. This is causing the largest problem
when a primary key and a foreign key differ just by a whitespace:
it is clear that they are intended to be the same, that is why we
always strip whitespaces.

Note that this is a warning:
* not an error because the feed may be still parsed;
* not an info because leading and trailing whitespaces are ambiguous
  and we want the feed providers to be aware of them.