feat: Do not treat NonAsciiOrNonPrintableCharNotice as an error (#728)
Issue #549

An error in RowParser stops validation for the whole feed. However, if
ID has non-printable or non-ASCII characters, the feed may be still
successfully parsed and served to the end user.

We are lowering priority of NonAsciiOrNonPrintableCharNotice to WARNING
and modify RowParser so that it checks severity level of its notices.

We also simplify the code of asId function and reduce amount of calls to
row.asString from 3 to 1.