feat: Report errors for new lines and carriage returns in field values
GTFS reference requires that field values do not contain new lines or
carriage returns (\n or \r).

RFC4180 CSV allows to spread a value between multiple lines if it is
surrounded by double quotes. But on practice, the feeds that have
multiline values are actually omitting the closing quote. That makes
different parsers interpret the file in different ways, so we need to
detect such ambiguous data in the validator.