feat: Handle final row that does not end with a newline
A CSV file may end with a row that contains only spaces and does not end
with a new line, so that row is actually empty. Univocity parser treats
that row as non-empty: it interprets it as a row with a single empty
column. It is unclear whether this is a bug or a feature.

RFC 4180 does not give a hint on how to treat such rows.

Here we handle such rows gracefully and show a warning since that CSV is
ambiguous.