feat: Split FieldParsingError into different classes
Format requirements for different fields are different, so that probably
does not make sense to group them all at the same error class.

Separate error classes may be placed in different sections of a
validation report that add details on the expected format.

Users of GTFS Validator may want to reduce the severity of certain
errors, e.g., validation of URLs and phone numbers.