Remove extra buffer - buffer on the Builder instead
This also has the effect of using Euclidean instead of Geodetic buffers, which saves execution time and likely memory, and doesn't seem to result in a level of error that causes problems. So the tradeoff of performance vs. accuracy seems acceptable. See https://github.com/MobilityData/gtfs-validator/pull/750#discussion_r578667817 for details.