docs: Add auto-gen Javadocs for Lists of @SequenceKey fields (#782)
The goal is to make the sorting of GTFS entities that have been annotated with @SequenceKey more transparent in List data structure. Currently for a developer writing a new rule it's not clear whether they are responsible for sorting entities like GtfsShape by shapePtSequence in the rule implementation or if those entities have already be sorted in preprocessing.

This change auto-generates Javadocs for methods that return Lists of entities to make it clear which have been sorted, and by what field. This should hopefully result in less uncertainty for developers implementing new rules and should avoid potential performance pitfalls of sorting being unnecessarily implemented within validation rules.