docs: Update README for executing v2 JAR
Previously we didn't have the v2 JAR for people to use, so we had IntelliJ instructions for setting up a Run configuration, etc.

This PR replaces the IntelliJ instructions with steps people can take to install Java and run the v2 release JAR.

It also adds a Rules section to the README, as this is what most users will be interested in (after running the validator).

I also moved "Build the code" towards the bottom near "Architecture", as fewer people are probably interested in that.