docs: Update new rules documentation
This PR attempts to simply the new rules documentation by taking the approach of providing an example, and then walking through sections of the example for both FileValidators and SingleEntityValidators, but always presenting FileValidators file to follow the initially introduced example of the TripUsabilityValidator. It also attempts to shorten text when possible and provide some additional context when needed.

It also fixes references to the SingleEntityValidator sections, some of which mixed references to FeedExpirationDateValidator and FeedServiceDateValidator.