feat: Propagate InterruptedException during feed loading
IBM wrote an article on the subject:
http://web.archive.org/web/20210201071452/https://www.ibm.com/developerworks/java/library/j-jtp05236/index.html

If throwing InterruptedException means that a method is a blocking method, then calling a blocking method means that your method is a blocking method too, and you should have a strategy for dealing with InterruptedException. Often the easiest strategy is to throw InterruptedException yourself, as shown in the putTask() and getTask() methods in Listing 1. Doing so makes your method responsive to interruption as well and often requires nothing more than adding InterruptedException to your throws clause.￼