feat: Treat --storage_directory as directory
The old implementation treated --storage_directory as a target file name
instead of a directory. We bring validator behaviour to its
documentation.

createFromUrl now receives a Path instead of String for maximum
flexibility.

ThreadInterruptedError is not used anymore and it was removed.

We use try-with-resources where possible.

We also rewrote unit tests for GtfsInput:

* tests for redirects are removed since we do not implement handling
  of redirects in Validator code;
* added a test for invalid URL to createFromUrlInMemory;
* made tests granular and avoided having multiple assertThat.