feat: Reduce severity of SameNameAndDescriptionForRouteNotice to WARNING
A route that has the same route_desc and route_short_name or
route_long_name may be still served to the end user, so this is not an
error. In fact, many consumers may just ignore route_desc completely.

Numerous feeds that are consumed by Google do have this notice and they
are considered valid.

We also update the comment for the notice.