feat: Validate that locations are reachable if a station has pathways
Validate that if a station has pathways, then each location is
reachable in both directions: there is a way to get to that location
from some entrance and a way to get from that location to some
entrance.

Notices are reported for platforms, boarding areas and generic nodes
but not for entrances or stations.

Notices are not reported for platforms that have boarding areas since
such platforms may not have incident pathways. Instead, notices are
reported for the boarding areas.