feat: Support Hierarchical Vehicle Types in StopTimeTravelSpeedValidator
HVTs are non-standard but widely used. The standard GTFS validator emits
a warning for an HVT but the original value can be retrieved via
GtfsRouteType.routeTypeValue().

There are feeds that run high-speed railway services. The default 200
km/h limit is too low for them, so it is worth to apply the usual 500
km/h limit for railway.