fix: Use `ImmutableMap.Builder.buildOrThrow()` instead of `build()`
ImmutableMap.Builder.build() throws if the same key was added to the builder twice. There were thousands of bugs because many people are not aware of that. The `build()` method is deprecated in Guava 31 and replaced by `buildOrThrow()` with the same behaviour.

See also: https://github.com/google/guava/commit/4b9c269f5cba1431101ffde0290a871769120316