docs: Remove dollar sign from commands
It's not actually part of the command you need to execute. IMHO the docs are better without it because it makes it easier to copy/paste if you're familiar with running things from the command line, and is less confusing if you're not as you'll get an error if you include the dollar sign in the command.