Generate registries of tables and validators (#1290)
Do not use the potentially unsafe ClassPath library that requires a lot of permissions and is problematic to run in a secure sandbox.

This is an API change.
Old code:
  validatorLoader = ValidatorLoader.createForDefaultPackage();
  feedLoader = new GtfsFeedLoader();

New code:
  validatorLoader = ValidatorLoader.createFromRegistry(new DefaultValidatorRegistry());
  feedLoader = new GtfsFeedLoader(new DefaultTableRegistry());

ValidatorLoader.createForPackages is removed. Create a custom implementation of GtfsValidatorRegistry and call ValidatorLoader.createFromRegistry.

We are still using ClassPath in NoticeSchemaGenerator and VersionResolver. They are not used for validation.