Do not generate a byTranslationKey() method implementation for multi-key schemas that do not have any translation record id annotations. (#1302)
The `@PrimaryKey(translationRecordIdType)` field is used to indicated which keys, if any, for a given GTFS entity can be used for performing `translations.txt` lookups via `byTranslationKey()` methods generated in the table container.  Currently, code in `TableContainerIndexGenerator` will generate an implementation for `byTranslationKey()` even if none of the @PrimaryKey fields have been marked as the translation RECORD_ID.  This can lead to some weird behavior where the method will return a value even if none should be returned or generate code that does not actually compile.

This PR fixes this behavior by no longer generating a `byTranslationKey()` implementation if the entity type does not support.  Instead, empty is returned.