Discover tables and validators in runtime using io.github.classgraph (#1311)
This effectively reverts commit 7fb105e2b759d21b638badbc312108bec9f43b15
where registries of tables and validators were generated by the
GtfsAnnotationProcessor. It is impossible to reliably generate a
registry for multiple classes since there is an orthogonality rule for
Java Annotation Processors [1]:

“The result of processing a given input is not a function of the presence or absence of other inputs (orthogonality).”

In other words, there is no way to guarantee that GtfsAnnotationProcessor
receives all schema and validator classes: it may receive only a portion of
them, so it cannot generate a global registry.

Discovery is implemented in a separate class ClassGraphDiscovery,
so that clients of the GTFS Validator may implement their own ways
to discover tables and validators. io.github.classgraph is just one
option, an alternative is com.google.common.reflect.

[1] https://docs.oracle.com/javase/7/docs/api/javax/annotation/processing/Processor.html