web/service/Dockerfile: fix permissions for application.properties, service acount json, & newrelic.yml
The gcloud-secrets CLI will create these files with 0600 permissions,
and does not document an option to change the mode. This patch uses the
chown/chmod CLI utilities rather than --chown and --chmod flags because
their support is not universal across docker versions or distributions
which do not include buildkit (the cloudbuild environment provides one
such distribution).