Changes logging for skipped validators to atInfo instead of atSevere.
We've started using the GTFS validator at Entur and this message is logged at too high a severity, with no way to disable to logging in the validator itself. If this change is not possible, then it would be great to have a way to opt out of it. Either by defining the validators required when running the validator or by setting a logging config in the validator.