fix: update TripAndShapeDistanceValidator to ignore shape with no shape_dist_traveled (#2018)
Some GTFS feeds do not provide shape_dist_traveled in shapes.txt
In this case, the value is set to 0 in the code, so sorting shapes based
on ShapeDistTraveled does not work and usually returns the first shape
row.

It leads to an error, the first shape of the shape is usually very far
away from the last stop of the trips following the shape.

We can assume the biggest shape_dist_traveled of a given shape should
never be 0, otherwise it means the value is not set. In this case, we
return early to ensure no validation error will be issued.

Fixes #2018